refactor: 상품 등록/수정 공통부분 리팩토링 - 등록, 수정 페이지에 서비스 함수들을 가져와서 submit 함수를 만들어서 공용등록폼에 전달해서 실행
